+++ crates/l2/Cargo.toml
@@ -13,7 +13,7 @@ tracing.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 ethereum-types.workspace = true
-ethrex-common = { workspace = true, features = ["websign"] }
+ethrex-common = { workspace = true, features = ["web3sign"] }
 ethrex-rlp.workspace = true
 ethrex-rpc.workspace = true
 ethrex-blockchain.workspace = true
